perf: remove manual caching / reuse stitches cache
instead of manually creating a style cache (to track which styles have already been inserted into the page), we can rely on stitches internal cache preventing duplicated styles, however we must *clear out* the cssRules that have already been inserted.

In order to restore these rules for the next server render, we can call `reset` on the provided config if we detect a render at the top of the context tree.